requirements.txt
# FastAPI Backend - LungCancerVR Simulator
# Version: 2.1 - Local/Open-Source

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.2.0

# Vector Database
chromadb==0.4.22

# Embeddings - BGE-M3 (multilingüe, soporta español)
sentence-transformers==2.3.1
transformers==4.36.2
# torch se instala por separado: pip install torch --index-url https://download.pytorch.org/whl/cpu

# RAG Orchestration
langchain==0.1.4
langchain-community==0.0.16

# PDF Processing (para indexación de documentos médicos)
pypdf==4.0.1
python-multipart==0.0.6

# HTTP Client (para Ollama)
httpx==0.26.0

# Database (PostgreSQL async)
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0

# Linting
flake8
